requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pipreqs==0.4.11
psycopg2-binary==2.9.5
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarg==0.1.9
